finish renaming to vagrant-unison2, current version is broken
