+++ lib/vagrant-unison2/command.rb
