+++ lib/vagrant-unison2/config.rb
