+++ lib/vagrant-unison2/errors.rb
