+++ lib/vagrant-unison2/shell_command.rb
