+++ lib/vagrant-unison2/ssh_command.rb
